Prohibit variadic signals/slots; change arity to paramCount
The use of variadic signals was a bit confusing and sort of defeated
the protection provided by the arity checks. I don't have any immediate
need for them, so I'm removing the functionality. Variadic slots have
also been removed. I could provide a flag to enable disabling arity
checks on a per connection basis, but I don't think that variadic
slots are going to be of much use without variadic signals.

The commit also renames arity to paramCount. It now represents the
number of parameters *excluding* the callback, which is implied
and obligatory in all situations.